requirements.txt
datetime~=4.4
pathlib~=1.0.1
smbus
pimoroni-bme280